netfs: Provide invalidatepage and releasepage calls
Provide default invalidatepage and releasepage calls.  These will need to
interact with invalidation correctly at some point.  They don't, however,
need to deal with the page fscache marks as those are no longer being used.

This also removes the byte-range dirty tracking in afs as that's intended
to be replaced.

Signed-off-by: David Howells <dhowells@redhat.com>